first_step_almost

+++ app/Http/Controllers/CommonController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Product;
+use App\Models\Order;
+use App\Models\OrderDetail;
 use App\Models\DeliveryCompany;
 use App\Models\Category;
 use App\Models\MainProduct;
@@ -30,26 +32,50 @@ class CommonController extends Controller
     }
     public function DeleteProduct($id)
     {
-        Product::where('id', $id)->delete();
+        $check = MainProduct::where('product_id',$id)->first();
+        if ($check) {
+            return Back()->with('error','You Can not delete this Product. It is used in product!');
+        } else {
+            Product::where('id', $id)->delete();
 
-        return redirect()->back();
+            return back()->with('error','Product Deleted Successfully!');
+        }
+        
     }
     public function DeleteCategory($id)
     {
-        Category::where('id', $id)->delete();
+        $check = Product::where('category_id',$id)->first();
+        if ($check) {
+            return Back()->with('error','You Can not delete this Category. It is used in product!');
+        } else {
+            Category::where('id', $id)->delete();
 
-        return redirect()->back();
+            return back()->with('error','Category Deleted Successfully!');
+        }
     }
     public function DeleteDeliveryCompany($id)
     {
-        DeliveryCompany::where('id', $id)->delete();
+            DeliveryCompany::where('id', $id)->delete();
 
-        return redirect()->back();
+            return back()->with('error','Delivery Company Deleted Successfully!');
     }
     public function DeleteSupplier($id)
     {
-        SupplierDetail::where('id', $id)->delete();
+        $check = Product::where('supplier_id',$id)->first();
+        if ($check) {
+            return Back()->with('error','You Can not delete this Supplier. It is used in product!');
+        } else {
+            SupplierDetail::where('id', $id)->delete();
 
-        return redirect()->back();
+            return back()->with('error','Supplier Deleted Successfully!');
+        }
+        
+    }
+
+    public function orders()
+    {
+        $data = Order::all();
+
+        return view('front.order-list',compact('data'));
     }
 }
